requirements.txt
azure-servicebus==0.50.2
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
msgpack==0.6.2
requests==2.24.0
retrying==1.3.3